+++ scout/adapter/mongo/filter.py
@@ -14,6 +14,7 @@ VARIANTS_TARGET_FROM_CATEGORY = {
     "cancer_sv": "variants.cancer_sv_variants",
     "snv": "variants.variants",
     "str": "variants.str_variants",
+    "fusion": "variants.fusion_variants",
 }