+++ scout/constants/file_types.py
@@ -5,6 +5,8 @@ FILE_TYPE_MAP = {
     "vcf_cancer_sv": {"category": "cancer_sv", "variant_type": "clinical"},
     "vcf_cancer_research": {"category": "cancer", "variant_type": "research"},
     "vcf_cancer_sv_research": {"category": "cancer_sv", "variant_type": "research"},
+    "vcf_fusion": {"category": "fusion", "variant_type": "clinical"},
+    "vcf_fusion_research": {"category": "fusion", "variant_type": "research"},
     "vcf_snv": {"category": "snv", "variant_type": "clinical"},
     "vcf_snv_research": {"category": "snv", "variant_type": "research"},
     "vcf_sv": {"category": "sv", "variant_type": "clinical"},