+++ scout/parse/variant/genotype.py
@@ -18,7 +18,7 @@ Uses 'DV' to describe number of paired ends that supports the event and
 
 """
 
-from typing import Dict, Tuple
+from typing import Dict, Optional, Tuple
 
 import cyvcf2
 
@@ -137,6 +137,8 @@ def parse_genotype(variant, ind, pos):
     gt_call["read_depth"] = get_read_depth(variant, pos, alt_depth, ref_depth)
     gt_call["alt_frequency"] = get_alt_frequency(variant, pos)
     gt_call["genotype_quality"] = int(variant.gt_quals[pos])
+    gt_call["ffpm"] = get_ffpm_info(variant, pos)
+    gt_call["split_read"] = split_read_alt
 
     return gt_call
 
@@ -180,6 +182,20 @@ def get_mei_reads(variant: cyvcf2.Variant, pos: Dict[str, int]) -> Tuple[int, ..
             pass
 
     return (spanning_ref, clip5_alt, clip3_alt)
+
+
+def get_ffpm_info(variant: cyvcf2.Variant, pos: Dict[str, int]) -> Optional[int]:
+    """Get FUSION caller read details from FORMAT tags.
+    Returns:
+        tuple(int, int, int) supporting_reads, split_reads, ffpm
+    """
+    # Fusion fragments per million total RNA-seq fragments
+    if "FFPM" in variant.FORMAT:
+        try:
+            values = variant.format("FFPM")[pos]
+            return int(values[0])
+        except ValueError as _ignore_error:
+            pass
 
 
 def get_paired_ends(variant, pos):