+++ tests/server/blueprints/variants/test_variants_views.py
@@ -254,6 +254,34 @@ def test_str_variants(app, institute_obj, case_obj):
         assert resp.status_code == 200
 
 
+def test_fusion_variants(
+    app, institute_obj, fusion_case_obj, fusion_variant_objs, one_fusion_variant
+):
+    """Test the page that displays a list of RNA fusion variants."""
+
+    # GIVEN an initialized app
+    with app.test_client() as client:
+        # GIVEN a valid user and institute
+        client.get(url_for("auto_login"))
+
+        # GIVEN a RNA fusion case present in the database
+        assert store.case_collection.insert_one(fusion_case_obj)
+
+        # GIVEN that the case has RNA fusion variants
+        store.variant_collection.insert_many([variant for variant in fusion_variant_objs])
+
+        # WHEN accessing the fusion variants page
+        resp = client.get(
+            url_for(
+                "variants.fusion_variants",
+                institute_id=institute_obj["internal_id"],
+                case_name=fusion_case_obj["display_name"],
+            )
+        )
+        # THEN it should return a page
+        assert resp.status_code == 200
+
+
 def test_mei_variants(app, institute_obj, case_obj):
     # GIVEN an initialized app
     # GIVEN a valid user and institute